add weighting function to sincronia

+++ htsim/datacenter/topology.h
@@ -29,7 +29,7 @@ public:
 
     virtual bool accommodate(const std::set<unsigned>&, const std::set<unsigned>&) {return false;};
 
-    virtual std::deque<uint64_t> bssi(std::unordered_map<Tensor *, double>) {return {};};
+    virtual std::deque<uint64_t> bssi(std::unordered_map<Tensor *, uint64_t>) {return {};};
 };
 
 #endif
